+++ tests/agent/actions/test_unknown.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import pytest
 
 from great_expectations_cloud.agent.actions.unknown import UnknownEventAction
+from great_expectations_cloud.agent.agent_warnings import GXAgentUserWarning
 from great_expectations_cloud.agent.models import UnknownEvent
-from great_expectations_cloud.agent.warnings import GXAgentUserWarning
 
 
 def test_unknown_throws_warning(mocker):
